Changed mgdraw so that printout is more general
There was an issue whereby some neutrons (at least one) was scored on the TPC file, but not the OD file which should contain ALL neutrons scored in the TPC file. I changed the assignment of the "kill" variable so that it is not assumed where it will be tagged first. It was assumed previously that it would be tagged in the OD part first.